更新bo对象

+++ src/main/java/org/edi/stocktask/bo/stockreport/IStockReportItem.java
@@ -1,7 +1,5 @@
 package org.edi.stocktask.bo.stockreport;
 
-import java.math.BigDecimal;
-
 /**
  * @author Fancy
  * @date 2018/5/25
@@ -23,7 +21,6 @@ public interface IStockReportItem {
     String getLineStatus();
 
     void setLineStatus(String  value);
-
 
     String getReference1();
 
@@ -101,9 +98,9 @@ public interface IStockReportItem {
 
     void setItemDescription(String value);
 
-    BigDecimal getQuantity();
+    Double getQuantity();
 
-    void setQuantity(BigDecimal value);
+    void setQuantity(Double value);
 
     String getInventoryUoM();
 
@@ -121,21 +118,21 @@ public interface IStockReportItem {
 
     void setServiceNumberManagement(String value);
 
-    BigDecimal getPrice();
+    Double getPrice();
 
-    void setPrice(BigDecimal value);
+    void setPrice(Double value);
 
     String getCurrency();
 
     void setCurrency(String value);
 
-    BigDecimal getCurrencyRate();
+    Double getCurrencyRate();
 
-    void setCurrencyRate(BigDecimal value);
+    void setCurrencyRate(Double value);
 
-    BigDecimal getLineTotal();
+    Double getLineTotal();
 
-    void setLineTotal(BigDecimal value);
+    void setLineTotal(Double value);
 
     String getFromWarehose();
